Fix security vulnerabilities and code quality issues from CodeRabbit review
Security fixes:
- Add Origin header validation for CSRF protection in PlaygroundController
- Fix XSS in form_submit_demo by escaping user input with ERB::Util.html_escape
- Fix ReDoS vulnerability with input length limit and non-greedy regex patterns
- Add sanitizeHtml() in studio_controller.js to prevent innerHTML XSS

Code quality fixes:
- Fix dynamic Tailwind classes in charts.rb with BAR_COLORS constant
- Fix btoa()/atob() Unicode issues with encodeURIComponent/decodeURIComponent
- Fix User.create bypass in modern_crud_demo.rb with proper ID assignment
- Fix type coercion in delete_user with .to_i
- Add bounds checking in torture_test_demo.rb select_coin
- Add column validation in move_card method

Also consolidates v2 playground into main playground (removes v2 files).
