--- playground_app/app/views/playground/_preview.html.erb
@@ -1,1 +1,0 @@
-<%= html.html_safe %>
