--- playground_app/app/views/playground_v2/index.html.erb
@@ -1,7 +1,0 @@
-<!-- Monaco Editor Scripts -->
-<script src="https://cdnjs.cloudflare.com/ajax/libs/monaco-editor/0.45.0/min/vs/loader.min.js"></script>
-<script>
-  require.config({ paths: { 'vs': 'https://cdnjs.cloudflare.com/ajax/libs/monaco-editor/0.45.0/min/vs' }});
-</script>
-
-<%= render @playground %>
